feat: combine plugin `weekYear` and `weekOfYear ` into `week`
