+++ src/plugins/index.ts
@@ -8,8 +8,7 @@ import minMaxPlugin from './minMax'
 import toArrayPlugin from './toArray'
 import toObjectPlugin from './toObject'
 import utcPlugin from './utc'
-import weekOfYearPlugin from './weekOfYear'
-import weekYearPlugin from './weekYear'
+import weekPlugin from './week'
 
 export {
   advancedFormatPlugin,
@@ -22,6 +21,5 @@ export {
   toArrayPlugin,
   toObjectPlugin,
   utcPlugin,
-  weekOfYearPlugin,
-  weekYearPlugin,
+  weekPlugin,
 }
