+++ test/plugins/weekYear.test.ts
@@ -2,10 +2,10 @@ import { esday } from 'esday'
 import moment from 'moment'
 import { describe, expect, it } from 'vitest'
 import localeEnUs from '~/locales/en-us'
-import { localePlugin, weekOfYearPlugin, weekYearPlugin } from '~/plugins'
+import { localePlugin, weekPlugin } from '~/plugins'
 
 // Extend esday with required plugins and locale
-esday.extend(localePlugin).extend(weekOfYearPlugin).extend(weekYearPlugin)
+esday.extend(localePlugin).extend(weekPlugin)
 esday.registerLocale(localeEnUs, 'en-us')
 esday.locale('en-us')
